Internally using a quality enum instead of numbers. Fixes #4
Software should be more interested in the *intentions* of the users,
not the necessarily the hard facts. Storing the qualities as numbers
breaks this spirit. The user didn't select a quality of 0.5, the user
selected Primitive. So we should store as Primitive, not 0.5. This
means when copying a set from one beacon to another, the user's
intentions follow instead of the exact values, allow the quality values
to be correctly computed for the destination beacon. It also allows
qualities to be adjusted just by opening a file and re-exporting.