Alternative for the blueprints problem.
In this version, set entries are duplicated. One entry for the actual
items, one entry for a blueprint. The weight of each will be divided by
the chance, which in theory, should do the trick. However, it does
consume additional entries from the min and max quantities. It's
possible to get both a usable item and its blueprint.