Backend support for engram control generation.
Thanks to this, Beacon will generate the most streamlined engram control code ever seen. For example, if bOnlyAllowSpecifiedEngrams is true and OverrideNamedEngramEntries is set to hide an engram, Beacon knows that is redundant so on generation, it'll omit the EngramHidden argument. If there are no arguments, it'll omit the OverrideNamedEngramEntries line entirely.