Improving rewriter to better match what the editors display.
In the past, including Beacon 1.4, the rewriter would ignore certain values that the user didn't change. For example, using stat multipliers would generate new multiplier lines but wouldn't remove the lines for other columns. So the user could get into a situation where Beacon shows a default multiplier, but their ini had a different config that Beacon would ignore because the multiplier was default. This change will rectify that.